+++ client/src/pages/database/index.tsx
@@ -1,5 +1,7 @@
 import React, { useState } from 'react';
 import { toast } from 'react-toastify';
+import { Listbox, Transition } from '@headlessui/react';
+import cx from 'classnames';
 import { Header } from '../../modules/layout/Header';
 import {
   useDatabaseByIdQuery,
@@ -8,9 +10,9 @@ import {
   useUnlinkDatabaseMutation,
   useUnlinkDatabaseLogsSubscription,
   useLinkDatabaseLogsSubscription,
+  RealTimeLog,
 } from '../../generated/graphql';
 import { useParams, Link } from 'react-router-dom';
-import Select from 'react-select';
 import {
   TabNav,
   TabNavLink,
@@ -26,10 +28,13 @@ export const Database = () => {
   const { id: databaseId } = useParams<{ id: string }>();
   const [isUnlinkModalOpen, setIsUnlinkModalOpen] = useState(false);
   const [isLinkModalOpen, setIsLinkModalOpen] = useState(false);
-  const [arrayOfUnlinkLogs, setArrayOfUnlinkLogs] = useState<string[]>([]);
-  const [arrayOfLinkLogs, setArrayOfLinkLogs] = useState<string[]>([]);
+  const [arrayOfUnlinkLogs, setArrayOfUnlinkLogs] = useState<RealTimeLog[]>([]);
+  const [arrayOfLinkLogs, setArrayOfLinkLogs] = useState<RealTimeLog[]>([]);
   const [appAboutToUnlink, setAppAboutToUnlink] = useState<string>();
   const [isTerminalVisible, setIsTerminalVisible] = useState(false);
+  const [processStatus, setProcessStatus] = useState<
+    'running' | 'notStarted' | 'finished'
+  >('notStarted');
   const [unlinkLoading, setUnlinkLoading] = useState(false);
   const [linkLoading, setLinkLoading] = useState(false);
 
@@ -60,21 +65,34 @@ export const Database = () => {
 
   useUnlinkDatabaseLogsSubscription({
     onSubscriptionData: (data) => {
-      const logsExist = data.subscriptionData.data?.unlinkDatabaseLogs?.[0];
-      if (logsExist)
+      const logsExist = data.subscriptionData.data?.unlinkDatabaseLogs;
+      if (logsExist) {
         setArrayOfUnlinkLogs((currentLogs) => {
           return [...currentLogs, logsExist];
         });
+        if (
+          logsExist.type === 'end:success' ||
+          logsExist.type === 'end:failure'
+        ) {
+          setProcessStatus('finished');
+        }
+      }
     },
   });
 
   useLinkDatabaseLogsSubscription({
     onSubscriptionData: (data) => {
-      const logsExist = data.subscriptionData.data?.linkDatabaseLogs?.[0];
+      const logsExist = data.subscriptionData.data?.linkDatabaseLogs;
       if (logsExist) {
         setArrayOfLinkLogs((currentLogs) => {
           return [...currentLogs, logsExist];
         });
+        if (
+          logsExist.type === 'end:success' ||
+          logsExist.type === 'end:failure'
+        ) {
+          setProcessStatus('finished');
+        }
       }
     },
   });
@@ -171,7 +189,7 @@ export const Database = () => {
         <div className="grid sm:grid-cols-1 md:grid-cols-2 lg:grid-cols-2 xl:grid-cols-2 gap-4 mt-10">
           <div>
             <h1 className="text-lg font-bold py-5">Database info</h1>
-            <div className="mt-3 bg-gray-100 shadow overflow-hidden rounded-lg border-b border-gray-200">
+            <div className="bg-gray-100 shadow overflow-hidden rounded-lg border-b border-gray-200">
               <table className="mt-4 mb-4 min-w-full bg-white">
                 <tbody className="text-gray-700">
                   <tr className="bg-gray-100">
@@ -221,18 +239,78 @@ export const Database = () => {
               <React.Fragment>
                 {notLinkedApps.length !== 0 ? (
                   <div>
-                    <Select
+                    <Listbox
+                      as="div"
                       value={selectedApp}
+                      //@ts-ignore
                       onChange={setSelectedApp}
-                      className="mt-3 w-80"
-                      options={appOptions}
-                      placeholder={selectedApp}
-                      isSearchable={false}
-                      aria-labelledby="app-select-dropdown"
-                      noOptionsMessage={() =>
-                        'All of your apps are already linked to this database'
-                      }
-                    />
+                    >
+                      {({ open }) => (
+                        <div className="relative w-80">
+                          <Listbox.Button className="cursor-default relative w-full rounded-md border border-gray-300 bg-white pl-3 pr-10 py-2 text-left focus:outline-none focus:shadow-outline-blue focus:border-blue-300 transition ease-in-out duration-150 sm:text-sm sm:leading-5">
+                            <span className="block truncate">
+                              {selectedApp.label}
+                            </span>
+                            <span className="absolute inset-y-0 right-0 flex items-center pr-2 pointer-events-none">
+                              <svg
+                                className="h-5 w-5 text-gray-400"
+                                viewBox="0 0 20 20"
+                                fill="none"
+                                stroke="currentColor"
+                              >
+                                <path
+                                  d="M7 7l3-3 3 3m0 6l-3 3-3-3"
+                                  strokeWidth="1.5"
+                                  strokeLinecap="round"
+                                  strokeLinejoin="round"
+                                />
+                              </svg>
+                            </span>
+                          </Listbox.Button>
+                          {open && (
+                            <Transition
+                              show={open}
+                              enter="transition ease-out duration-100"
+                              enterFrom="transform opacity-0 scale-95"
+                              enterTo="transform opacity-100 scale-100"
+                              leave="transition ease-in duration-75"
+                              leaveFrom="transform opacity-100 scale-100"
+                              leaveTo="transform opacity-0 scale-95"
+                              className="absolute mt-1 w-full rounded-md bg-white shadow-lg z-10"
+                            >
+                              <Listbox.Options
+                                static
+                                className="max-h-60 rounded-md py-1 text-base leading-6 shadow-xs overflow-auto focus:outline-none sm:text-sm sm:leading-5"
+                              >
+                                {appOptions.map(
+                                  (app) =>
+                                    app.value.id !== selectedApp.value.id && (
+                                      <Listbox.Option
+                                        key={appOptions.indexOf(app)}
+                                        value={app as any}
+                                      >
+                                        {({ active }) => (
+                                          <div
+                                            className={cx(
+                                              'cursor-default select-none relative py-2 px-4',
+                                              {
+                                                'bg-gray-200': active,
+                                                'bg-white text-black': !active,
+                                              }
+                                            )}
+                                          >
+                                            {app.label}
+                                          </div>
+                                        )}
+                                      </Listbox.Option>
+                                    )
+                                )}
+                              </Listbox.Options>
+                            </Transition>
+                          )}
+                        </div>
+                      )}
+                    </Listbox>
 
                     {databaseLinkError && (
                       <p className="text-red-500 text-sm font-semibold">
@@ -263,18 +341,18 @@ export const Database = () => {
                                 Linking <b>{selectedApp.value.name}</b> with{' '}
                                 <b>{database.name}</b>!
                               </p>
+                              <p className="text-gray-500 mb-2">
+                                Linking process usually takes a couple of
+                                minutes. Breathe in, breathe out, logs are about
+                                to appear below:
+                              </p>
                               <Terminal className={'w-6/6'}>
-                                <p className="text-green-400 mb-2">
-                                  Linking process usually takes a couple of
-                                  minutes. Breathe in, breathe out, logs are
-                                  about to appear below:
-                                </p>
                                 {arrayOfLinkLogs.map((log) => (
                                   <p
                                     key={arrayOfLinkLogs.indexOf(log)}
                                     className="text-s leading-5"
                                   >
-                                    {log}
+                                    {log.message}
                                   </p>
                                 ))}
                               </Terminal>
@@ -288,18 +366,21 @@ export const Database = () => {
                           )}
                         </ModalDescription>
                         <ModalButton
-                          ctaFn={() =>
-                            handleConnect(databaseId, selectedApp.value.id)
-                          }
+                          ctaFn={() => {
+                            setProcessStatus('running');
+                            handleConnect(databaseId, selectedApp.value.id);
+                          }}
                           ctaText={'Link'}
                           otherButtonText={'Cancel'}
                           isCtaLoading={isTerminalVisible ? false : linkLoading}
                           isCtaDisabled={isTerminalVisible}
+                          isOtherButtonDisabled={processStatus === 'running'}
                           closeModal={() => {
                             setIsLinkModalOpen(false);
                             refetch({ databaseId });
                             setLinkLoading(false);
                             setIsTerminalVisible(false);
+                            setProcessStatus('notStarted');
                           }}
                         />
                       </Modal>
@@ -362,18 +443,18 @@ export const Database = () => {
                                     Unlinking <b>{database.name}</b> from{' '}
                                     <b>{appAboutToUnlink}</b>!
                                   </p>
+                                  <p className="text-gray-500 mb-2">
+                                    Unlinking process usually takes a couple of
+                                    minutes. Breathe in, breathe out, logs are
+                                    about to appear below:
+                                  </p>
                                   <Terminal className={'w-6/6'}>
-                                    <p className="text-green-400 mb-2">
-                                      Unlinking process usually takes a couple
-                                      of minutes. Breathe in, breathe out, logs
-                                      are about to appear below:
-                                    </p>
                                     {arrayOfUnlinkLogs.map((log) => (
                                       <p
                                         key={arrayOfUnlinkLogs.indexOf(log)}
                                         className="text-s leading-5"
                                       >
-                                        {log}
+                                        {log.message}
                                       </p>
                                     ))}
                                   </Terminal>
@@ -387,11 +468,17 @@ export const Database = () => {
                               )}
                             </ModalDescription>
                             <ModalButton
-                              ctaFn={() => handleUnlink(database.id, app.id)}
+                              ctaFn={() => {
+                                setProcessStatus('running');
+                                handleUnlink(database.id, app.id);
+                              }}
                               ctaText={'Unlink'}
                               otherButtonText={'Cancel'}
                               isCtaLoading={
                                 isTerminalVisible ? false : unlinkLoading
+                              }
+                              isOtherButtonDisabled={
+                                processStatus === 'running'
                               }
                               isCtaDisabled={isTerminalVisible}
                               closeModal={() => {
@@ -400,6 +487,7 @@ export const Database = () => {
                                 setUnlinkLoading(false);
                                 setIsTerminalVisible(false);
                                 setAppAboutToUnlink('');
+                                setProcessStatus('notStarted');
                               }}
                             />
                           </Modal>