+++ server/src/graphql/mutations/index.ts
@@ -7,6 +7,9 @@ import { destroyApp } from './destroyApp';
 import { destroyDatabase } from './destroyDatabase';
 import { linkDatabase } from './linkDatabase';
 import { unlinkDatabase } from './unlinkDatabase';
+import { addAppProxyPort } from './addAppProxyPort';
+import { removeAppProxyPort } from './removeAppProxyPort';
+import { restartApp } from './restartApp';
 
 export const mutations = {
   loginWithGithub,
@@ -15,7 +18,10 @@ export const mutations = {
   setEnvVar,
   unsetEnvVar,
   destroyApp,
+  restartApp,
   destroyDatabase,
   linkDatabase,
   unlinkDatabase,
+  addAppProxyPort,
+  removeAppProxyPort,
 };