+++ server/src/lib/dokku/index.ts
@@ -15,6 +15,10 @@ import { links as databaseLinks } from './plugin/database/links';
 import { link } from './plugin/database/link';
 import { unlink } from './plugin/database/unlink';
 import { create as createDatabase } from './plugin/database/create';
+import { proxyPorts } from './proxy/ports';
+import { proxyPortsAdd } from './proxy/portsAdd';
+import { proxyPortsRemove } from './proxy/portsRemove';
+import { restart } from './process/restart';
 
 export const dokku = {
   apps: { create, logs, destroy, list: appList },
@@ -29,5 +33,13 @@ export const dokku = {
     create: createDatabase,
     link,
     unlink,
+  },
+  proxy: {
+    ports: proxyPorts,
+    portsAdd: proxyPortsAdd,
+    portsRemove: proxyPortsRemove,
+  },
+  process: {
+    restart,
   },
 };